auth-service/requirements.txt
fastapi==0.111.0
redis==5.0.4
pydantic-settings==2.8.0
sqlalchemy[asyncio]==2.0.39
asyncpg==0.30.0
Werkzeug==3.1.3
alembic==1.15.1
psycopg2==2.9.10
PyJWT==2.10.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
Authlib==1.5.2
itsdangerous==2.2.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-instrumentation-redis==0.41b0
fastapi-limiter==0.1.6